% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PER_2015_magn}
\alias{PER_2015_magn}
\title{Visual magnitude observations of Perseids from 2015}
\format{
A list with the same structure as returned by \link{load_vmdb_magnitudes}.
}
\description{
Visual magnitude observations of the Perseid shower from 2015.
}
\details{
\code{PER_2015_magn} are magnitude observations loaded with
\link{load_vmdb_magnitudes}.
}
\seealso{
\link{load_vmdb}
}
\keyword{data}
