% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PER_2015_rates}
\alias{PER_2015_rates}
\title{Visual rate observations of Perseids from 2015}
\description{
Visual rate and magnitude observations of the Perseid shower from 2015.
}
\details{
\code{PER_2015_rates} are rate observations loaded with \link{load_vmdb_rates}.
}
\seealso{
\link{load_vmdb}
}
\keyword{data}
