% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vismeteor-package}
\alias{vismeteor}
\alias{vismeteor-package}
\title{vismeteor: Analysis of Visual Meteor Data}
\description{
Provides a suite of analytical functionalities to process and analyze visual meteor observations from the Visual Meteor Database of the International Meteor Organization \url{https://www.imo.net/}.
}
\details{
The data used in this package can created and provided by
\href{https://pypi.org/project/imo-vmdb/}{imo-vmdb}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jankorichter/vismeteor}
  \item Report bugs at \url{https://github.com/jankorichter/vismeteor/issues}
}

}
\author{
\strong{Maintainer}: Janko Richter \email{janko@richtej.de}

}
